requirement.txt
Django>=4.2
requests
pymongo
python-dotenv